// src/app/models/PollingMember.ts
export class PollingMember{
  id?: number;
  personId?: number;
  parentId?: number;
  personName?: string;
  parentName?: string;
  remark?: string;
  email?: string;
  personTypeName?: string;
  age?: number;
  gender?: string;
  mobile1?: string;
  mobile2?: string;
  voterId?: string;
  assemblyName?: string;
  pollingNumber?: number;
}